0.7 with async

+++ src/types/option_value.rs
@@ -259,7 +259,7 @@ pub struct OptionValueInteger {
   #[serde(rename(serialize = "@extra", deserialize = "@extra"))]
   extra: Option<String>,
   /// The value of the option
-  value: i64,
+  value: isize,
   
 }
 
@@ -283,7 +283,7 @@ impl OptionValueInteger {
     RTDOptionValueIntegerBuilder { inner }
   }
 
-  pub fn value(&self) -> i64 { self.value }
+  pub fn value(&self) -> isize { self.value }
 
 }
 
@@ -296,7 +296,7 @@ impl RTDOptionValueIntegerBuilder {
   pub fn build(&self) -> OptionValueInteger { self.inner.clone() }
 
    
-  pub fn value(&mut self, value: i64) -> &mut Self {
+  pub fn value(&mut self, value: isize) -> &mut Self {
     self.inner.value = value;
     self
   }
